+++ Changelog.md
@@ -118,6 +118,9 @@
     1. Snapping on by default, off holding Shift. Improved snapping.
         * [Pull request #1304](https://bitbucket.org/osrf/gazebo/pull-request/1304)
 
+    1. Snap walls to length increments, moved scale to SegmentItem and added Get/SetScale, added SegmentItem::SnapAngle and SegmentItem::SnapLength.
+        * [Pull request #1311](https://bitbucket.org/osrf/gazebo/pull-request/1311)
+
     1. Added EditorItem::SetHighlighted.
         * [Pull request #1308](https://bitbucket.org/osrf/gazebo/pull-request/1308)